+++ dev/FoodForThought/src/features/restaurants/screens/restaurants.screen.js
@@ -5,10 +5,13 @@ import styled from "styled-components/native";
 
 import { SafeArea } from "../../../utils/safe-area.component";
 import { FavoritesBar } from "../../../components/favorites/favorites-bar.component";
+import { LocationContext } from "../../../services/location/location.context";
 import { RestaurantsContext } from "../../../services/restaurants/restaurants.context";
 import { FavoritesContext } from "../../../services/favorites/favorites.context";
 import { Search } from "../components/search.component";
 import { RestaurantInfoCard } from "../components/restaurant-info-card.component";
+import { Typography } from "../../../components/typography/text.component";
+import { FadeInView } from "../../../components/animations/fade.animation";
 
 const LoadingContainer = styled(View)`
   position: absolute;
@@ -26,9 +29,12 @@ const RestaurantList = styled(FlatList).attrs({
 })``;
 
 export const RestaurantScreen = ({ navigation }) => {
+  const { error, locationError } = useContext(LocationContext);
   const { isLoading, restaurants } = useContext(RestaurantsContext);
   const { favorites } = useContext(FavoritesContext);
   const [isToggled, setIsToggled] = useState(false);
+
+  const hasError = !!error || !!locationError;
 
   return (
     <SafeArea>
@@ -37,30 +43,34 @@ export const RestaurantScreen = ({ navigation }) => {
           <Loading size={100} animating={true} color={"#4c6a14"} />
         </LoadingContainer>
       )}
-      <Search
-        isFavoritesToggled={isToggled}
-        onFavoritesToggle={() => setIsToggled(!isToggled)}
-      />
-      {isToggled ? (
-        <FavoritesBar favorites={favorites} onNavigate={navigation.navigate} />
-      ) : null}
-      <RestaurantList
-        data={restaurants}
-        renderItem={({ item }) => {
-          return (
-            <TouchableOpacity
-              onPress={() =>
-                navigation.navigate("RestaurantDetail", {
-                  restaurant: item,
-                })
-              }
-            >
-              <RestaurantInfoCard restaurant={item} />
-            </TouchableOpacity>
-          );
-        }}
-        keyExtractor={(item) => item.name}
-      />
+      <Search isFavoritesToggled={isToggled} onFavoritesToggle={() => setIsToggled(!isToggled)} />
+      {isToggled ? <FavoritesBar favorites={favorites} onNavigate={navigation.navigate} /> : null}
+      {hasError && (
+        <View>
+          <Typography variant="error">Something Went Wrong Retrieving Data</Typography>
+        </View>
+      )}
+      {!hasError && (
+        <RestaurantList
+          data={restaurants}
+          renderItem={({ item }) => {
+            return (
+              <TouchableOpacity
+                onPress={() =>
+                  navigation.navigate("RestaurantDetail", {
+                    restaurant: item,
+                  })
+                }
+              >
+                <FadeInView>
+                  <RestaurantInfoCard restaurant={item} />
+                </FadeInView>
+              </TouchableOpacity>
+            );
+          }}
+          keyExtractor={(item) => item.name}
+        />
+      )}
     </SafeArea>
   );
 };