+++ dev/FoodForThought/src/services/authentication/authentication.context.js
@@ -1,0 +1,82 @@
+import React, { useState, createContext, useRef } from "react";
+import {
+  signOut,
+  createUserWithEmailAndPassword,
+  onAuthStateChanged,
+  getAuth,
+} from "firebase/auth";
+
+import { loginRequest } from "./authentication.service";
+
+export const AuthenticationContext = createContext();
+
+export const AuthenticationContextProvider = ({ children }) => {
+  const [isLoading, setIsLoading] = useState(false);
+  const [user, setUser] = useState(null);
+  const [error, setError] = useState([]);
+  const auth = useRef(getAuth()).current;
+
+  onAuthStateChanged(auth, (usr) => {
+    if (usr) {
+      setUser(usr);
+      setIsLoading(false);
+    } else {
+      setIsLoading(false);
+    }
+  });
+
+  const onLogin = (email, password) => {
+    setIsLoading(true);
+    loginRequest(auth, email, password)
+      .then((u) => {
+        setUser(u);
+        setIsLoading(false);
+      })
+      .catch((err) => {
+        setIsLoading(false);
+        setError(err.toString());
+      });
+  };
+
+  const onRegister = (email, password, repeatedPassword) => {
+    setIsLoading(true);
+    if (password !== repeatedPassword) {
+      setError("Error: Passwords do not match.");
+      return;
+    }
+    createUserWithEmailAndPassword(auth, email, password)
+      .then((u) => {
+        setUser(u);
+        setIsLoading(false);
+      })
+      .catch((err) => {
+        setIsLoading(false);
+        setError(err.toString());
+      });
+  };
+
+  const onLogout = () => {
+    signOut(auth)
+      .then(() => {
+        setUser(null);
+        setError(null);
+      })
+      .catch((err) => console.error(err));
+  };
+
+  return (
+    <AuthenticationContext.Provider
+      value={{
+        isAuthenticated: !!user,
+        user,
+        isLoading,
+        error,
+        onLogin,
+        onRegister,
+        onLogout,
+      }}
+    >
+      {children}
+    </AuthenticationContext.Provider>
+  );
+};